tests/submission/W3C/domhtml/HTMLDocument09.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/dom.html#dom-document-links" rel="help">
<link rel="author" title="NIST">
</head>
<body id="TEST-BODY"> 
<h1>This is a Document</h1>
<p>This is a paragraph.</p>
<img alt="Yellow image" height="50" src="/support/images/yellow-rectangle.png" width="100">
<p class="copyright"><a href="http://www.csail.mit.edu/"><acronym title="Massachusetts Institute of Technology">MIT</acronym></a>, <a href="http://www.ercim.eu/"><acronym title="European Research Consortium for Informatics and Mathematics">ERCIM</acronym></a>,
<a href="http://www.keio.ac.jp/">Keio</a></p>


<script>
var domtest = async_test("The links attribute returns a collection of all AREA and A elements     in a document with a value for the href attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vlinks;
      var vlength;
      var doc;
      doc = window.document;
      vlinks = doc.links;

      vlength = vlinks.length;

      assert_equals(vlength,3);
       
    domtest.done();
}));

</script>
<p>The links attribute returns a collection of all AREA and A elements
    in a document with a value for the href attribute.

    Retrieve the links attribute from the document and examine its value.</p>
<div id="log"></div>
</body></html>
